Add NewTabPageFavoritesClient
